Move `QuintExprConverter` back into `Quint`
